chore: resolve circular dependency between Fn and FnGenerated
