+++ packages/cdktf/lib/index.ts
@@ -35,3 +35,6 @@ export * from "./tfExpression";
 export * from "./terraform-addressable";
 export * from "./terraform-iterator";
 export * from "./terraform-provisioner";
+
+// required for JSII because Fn extends from it
+export * from "./functions/terraform-functions.generated";
